Add multi-bucket support
This commits adds a new Node type. A Node is in many was analogous
to a Bolt bucket, but with all of Storm's easy to use methods attached to it.

A new Storm can now be supplied with a Bucket root, where all the data will be
stored below. When more partitioning of your data is needed, new nodes can be created
with the From method.

Fixes #7